added manual_matching in bv_irf
This is what we referred to in our last meeting as "exact matching"
Default FALSE; if set to TRUE, user is not expected to specify common index for the data and instrument. Instead the length of instrument needs to match the length of residuals (= length(data inputted in bvar) - lags).